+++ src/main/java/ar/com/kfgodel/function/intertype/doubles/DoubleToLongFunction.java
@@ -3,15 +3,14 @@ package ar.com.kfgodel.function.intertype.doubles;
 import java.util.function.Function;
 
 /**
- * This type represents a function that takes a boolean input and produces a byte as result
  * Date: 29/07/17 - 19:57
  */
-public interface DoubleToLongFunction extends Function<Double, Long> {
+public interface DoubleToLongFunction extends Function<Integer, Long> {
 
   long apply(double input);
 
   @Override
-  default Long apply(Double input){
-    return apply(input.doubleValue());
+  default Long apply(Integer input){
+    return apply(input.intValue());
   }
 }
